requirements.txt
langgraph==0.2.35
langchain-core==0.3.13
scikit-learn==1.5.2
numpy==2.1.1
rich==13.7.1
python-dotenv==1.0.1
streamlit==1.37.1
pillow==10.4.0
